Basically the whole project is ready
The whole project is a very simple script that wraps around Ghostty as
to allow Gnome's .desktop icons to use it to launch terminal-based
applications in it.

It basically strips the first argument passed by these launchers and
pass the rest to 'ghostty -e'.

An AUR package was created basically so that I don't have to store this
shit locally and can install it more easily on a future Arch install.
This package also attempts to make Ghostty the default terminal emulator
in all the ways I've found on Stack Overflow, forums, tutorials, etc.